refactor : 영화 도메인 모델의 포스터 이미지 프로퍼티 타입 수정

+++ app/src/main/java/woowacourse/movie/view/movie/viewholder/MovieViewHolder.kt
@@ -19,7 +19,7 @@ class MovieViewHolder(
     private val bookingBtn = itemView.findViewById<Button>(R.id.btn_booking)
 
     fun bind(item: Movie) {
-        moviePoster.setImageResource(item.poster)
+        moviePoster.setImageResource(item.poster.toInt())
         movieTitle.text = item.title
         movieReleaseDate.text =
             itemView.context.getString(R.string.text_date_period).format(
